[measureimagequality]: (5) Use pydantic measurements model, move funcitonal code to library/functions/measurement
